Cleanup the organization of widget modules and add documentation
